Added a test onboarding level and other minor additions
In this build I have mapped out a "test onboarding level" which will be used tomorrow (30/03/23) during our playtest to show how the level will play out and where onboarding will take place. Also I have set up a MeteorNavBP and corresponding foliage actor so that we can use meteors that have the nav mesh obstacle bounds attached to them within the foliage tool therefore saving us a lot of time in the future )note it is currently only set up for meteor1. Also added a basic tool tip feature for use in the playtest but can be expanded on.